Restore and reimplement `platform` handlers with improved struct definition and request validation
