--- internal/platform/delete.go
@@ -1,32 +1,0 @@
-package platform
-
-import (
-	"context"
-	"errors"
-	"net/http"
-	"products/internal"
-	"time"
-
-	"github.com/jackc/pgx/v5"
-)
-
-func (h *platformHandler) DeletePlatform(w http.ResponseWriter, r *http.Request) {
-	id, ok := internal.GetIntFromRequestPath("id", r)
-	if !ok {
-		http.Error(w, "Invalid platform ID", http.StatusBadRequest)
-		return
-	}
-	contextWithTimeOut, cancel := context.WithTimeout(r.Context(), 5*time.Second)
-	defer cancel()
-	_, err := h.queries.DeletePlatform(contextWithTimeOut, id)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			http.Error(w, "Platform not found", http.StatusNotFound)
-			return
-		}
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusNoContent)
-}
